Use MUnit in HeaderDecoderSuite
